feat(contracts): update data model and contract controller
Refactored the AddContractRequest and Contract classes. Updated fields from enums to strings, aligning with changes in the database schema. Altered corresponding methods for contract filtering within the ContractController. Added additional methods in ContractRepo for retrieving suggestions based on contract attributes such as company, currency, and contract number. Updated the UpdateContractRequest record to match the changes made to the AddContractRequest class.

Added method stubs for handling contract files as well as todos for future implementations relating to contract files. Adjusted the server port configuration in the application.yml file from 9000 to 3000.